# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backpackers API",
    "description": "국립공원·캠핑장 정보 플랫폼 API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "{{base_url}}",
      "description": "Lightsail (prod)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "your-password" },
          "name": { "type": "string", "example": "홍길동" }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "your-password" }
        }
      },
      "GoogleVerifyRequest": {
        "type": "object",
        "required": ["id_token"],
        "properties": {
          "id_token": { "type": "string", "example": "<Google ID Token>" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "user": { "$ref": "#/components/schemas/UserResponse" }
        }
      },
      "LocationSchema": {
        "type": "object",
        "properties": {
          "lat": { "type": "number", "example": 38.1195 },
          "lng": { "type": "number", "example": 128.4656 },
          "address": { "type": "string", "example": "강원도 속초시 설악산로 1234" },
          "region": { "type": "string", "example": "강원" }
        }
      },
      "CampingSpotSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "location": { "$ref": "#/components/schemas/LocationSchema" },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "tags": { "type": "array", "items": { "type": "string" } },
          "rating": { "type": "number" },
          "review_count": { "type": "integer" },
          "status": { "type": "string", "enum": ["verified", "pending", "rejected"] },
          "thumbnail_image": { "type": "string", "nullable": true }
        }
      },
      "CampingSpotDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "location": { "$ref": "#/components/schemas/LocationSchema" },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "tags": { "type": "array", "items": { "type": "string" } },
          "images": { "type": "array", "items": { "type": "string" } },
          "source": { "type": "string", "enum": ["public_data", "user_report"] },
          "status": { "type": "string", "enum": ["verified", "pending", "rejected"] },
          "rating": { "type": "number" },
          "review_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "PaginatedSpotsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CampingSpotSummary" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
      "SpotReportCreate": {
        "type": "object",
        "required": ["name", "description", "lat", "lng", "address"],
        "properties": {
          "name": { "type": "string", "example": "비밀 캠핑장" },
          "description": { "type": "string", "example": "아무도 모르는 숨겨진 캠핑장" },
          "lat": { "type": "number", "example": 37.5665 },
          "lng": { "type": "number", "example": 126.9780 },
          "address": { "type": "string", "example": "서울특별시 중구 ..." },
          "amenities": {
            "type": "array",
            "items": { "type": "string", "enum": ["toilet", "parking", "water", "shower", "fire_pit", "trash"] },
            "example": ["toilet", "parking"]
          },
          "images": { "type": "array", "items": { "type": "string" }, "example": [] },
          "reporter_contact": { "type": "string", "nullable": true, "example": "010-1234-5678" }
        }
      },
      "SpotReportResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "lat": { "type": "number" },
          "lng": { "type": "number" },
          "address": { "type": "string" },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "images": { "type": "array", "items": { "type": "string" } },
          "reporter_contact": { "type": "string", "nullable": true },
          "status": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
      "SpotSummary": {
        "type": "object",
        "properties": {
          "uid": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "example": "설악산 오토캠핑장" },
          "address": { "type": "string", "nullable": true, "example": "강원특별자치도 속초시 설악산로 1234" },
          "region_province": { "type": "string", "nullable": true, "example": "강원특별자치도" },
          "region_city": { "type": "string", "nullable": true, "example": "속초시" },
          "latitude": { "type": "number", "nullable": true, "example": 38.1195 },
          "longitude": { "type": "number", "nullable": true, "example": 128.4656 },
          "category": { "type": "array", "nullable": true, "items": { "type": "string" }, "example": ["일반야영장"] },
          "amenities": { "type": "array", "nullable": true, "items": { "type": "string" }, "example": ["주차장", "화장실", "샤워장"] },
          "themes": { "type": "array", "nullable": true, "items": { "type": "string" }, "example": ["숲속", "계곡"] },
          "is_pet_allowed": { "type": "boolean", "nullable": true },
          "is_fee_required": { "type": "boolean", "nullable": true },
          "rating_avg": { "type": "number", "example": 4.5 },
          "review_count": { "type": "integer", "example": 12 }
        }
      },
      "SpotDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/SpotSummary" },
          {
            "type": "object",
            "properties": {
              "description": { "type": "string", "nullable": true },
              "tagline": { "type": "string", "nullable": true },
              "features": { "type": "string", "nullable": true },
              "phone": { "type": "string", "nullable": true },
              "website_url": { "type": "string", "nullable": true },
              "booking_url": { "type": "string", "nullable": true },
              "nearby_facilities": { "type": "array", "nullable": true, "items": { "type": "string" } },
              "fire_pit_type": { "type": "string", "nullable": true },
              "camp_sight_type": { "type": "string", "nullable": true },
              "has_equipment_rental": { "type": "array", "nullable": true, "items": { "type": "string" } },
              "total_area_m2": { "type": "number", "nullable": true },
              "unit_count": { "type": "integer", "nullable": true },
              "created_at": { "type": "string", "format": "date-time" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
      "SpotSearchResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SpotSummary" } },
          "total": { "type": "integer", "example": 150 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total_pages": { "type": "integer", "example": 8 }
        }
      },
      "RegionCity": {
        "type": "object",
        "properties": {
          "province": { "type": "string", "example": "강원특별자치도" },
          "cities": { "type": "array", "items": { "type": "string" }, "example": ["속초시", "강릉시", "춘천시"] }
        }
      },
      "RegionListResponse": {
        "type": "object",
        "properties": {
          "regions": { "type": "array", "items": { "$ref": "#/components/schemas/RegionCity" } }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "서버 상태 확인",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "environment": { "type": "string", "example": "prod" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "이메일 회원가입",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserRegister" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "400": {
            "description": "이미 사용 중인 이메일",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "이메일 로그인",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserLogin" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": {
            "description": "이메일 또는 비밀번호 불일치",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/google/verify": {
      "post": {
        "tags": ["Auth"],
        "summary": "Google 소셜 로그인",
        "description": "프론트엔드가 Google Sign-In SDK로 획득한 ID Token을 전달하면 백엔드에서 검증 후 앱 토큰 발급. 신규 사용자는 자동 가입.",
        "operationId": "googleVerify",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GoogleVerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "인증 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "400": {
            "description": "유효하지 않은 Google 토큰",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/spots": {
      "get": {
        "tags": ["Spots"],
        "summary": "캠핑장 목록 조회",
        "operationId": "listSpots",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "이름·설명·태그 검색",
            "schema": { "type": "string" }
          },
          {
            "name": "region",
            "in": "query",
            "description": "지역 필터",
            "schema": {
              "type": "string",
              "enum": ["서울", "경기", "강원", "충북", "충남", "전북", "전남", "경북", "경남", "제주"]
            }
          },
          {
            "name": "amenities",
            "in": "query",
            "description": "편의시설 필터 (복수 선택 가능)",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["toilet", "parking", "water", "shower", "fire_pit", "trash"]
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호 (1부터)",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 항목 수 (최대 100)",
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedSpotsResponse" }
              }
            }
          }
        }
      }
    },
    "/spots/{spot_id}": {
      "get": {
        "tags": ["Spots"],
        "summary": "캠핑장 상세 조회",
        "operationId": "getSpot",
        "parameters": [
          {
            "name": "spot_id",
            "in": "path",
            "required": true,
            "description": "캠핑장 UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampingSpotDetail" }
              }
            }
          },
          "404": {
            "description": "캠핑장을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/spots/report": {
      "post": {
        "tags": ["Spots"],
        "summary": "캠핑장 제보",
        "operationId": "reportSpot",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SpotReportCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "제보 접수 완료",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SpotReportResponse" }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/spots/regions": {
      "get": {
        "tags": ["Spots"],
        "summary": "지역 목록 조회",
        "description": "지역 필터 드롭다운 구성용. 도(province) → 시/군/구(cities) 계층 구조로 반환. 인증 불필요.",
        "operationId": "listRegions",
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegionListResponse" },
                "example": {
                  "regions": [
                    { "province": "강원특별자치도", "cities": ["강릉시", "속초시", "춘천시"] },
                    { "province": "경기도", "cities": ["가평군", "양평군"] }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}